extract TimestampSelector component out of menu bar
